Extract OssIndexDocument classes to their own files
